Small refactor, added audio
