+++ src/minion.rs
@@ -1,7 +1,7 @@
-use std::borrow::BorrowMut;
+
 use bevy::prelude::*;
 
-use crate::{ player::Player, Velocity, Bandit, change_state, get_nearest_entity, avoidance, BindUi, HasUi, Attack, Health, Damage, Attacks };
+use crate::{ player::Player, Velocity, change_state, get_nearest_entity, BindUi, HasUi, Attack, Health, Damage, Attacks, components::Bandit };
 
 #[derive(Component)]
 pub struct Minion;
@@ -22,7 +22,6 @@ impl Plugin for MinionPlugin {
     fn build(&self, app: &mut App){
         app      
          .add_systems(Update, minion_follow_player)
-         .add_systems(Update, minion_avoidance)
          .add_systems(Update, check_enemy_range)
          .add_systems(Update, move_to_enemy)
          .add_systems(Update, attack_enemy)
@@ -36,7 +35,6 @@ pub struct SpawnMinion(pub Vec3);
 
 fn on_spawn_minions(mut commands: Commands, asset_server: Res<AssetServer>, mut writer: EventWriter<BindUi>, mut reader:EventReader<SpawnMinion>){
     for evt in  reader.iter(){
-        println!("SPAWN");
         let texture = asset_server.load("knight.png");
         let entity = commands.spawn((SpriteBundle {
             sprite: Sprite {
@@ -85,39 +83,30 @@ pub fn minion_movement(
     }
 }
 
-
-pub fn minion_avoidance(mut minion_query: Query<(&mut Transform, &mut Velocity), (With<Minion>, Without<Player>)>){
-    let mut combinations = minion_query.iter_combinations_mut();
-    while let Some([mut t1, t2]) = combinations.fetch_next(){
-        if t1.0.translation.distance(t2.0.translation) > 50.0 { continue;}
-        let t1_position = t1.0.translation;
-        t1.borrow_mut().1.0 -= avoidance(t2.0.translation, t1_position);
-    }
-}
-
 pub fn move_to_enemy(
     mut commands: Commands,
     mut minion_query: Query<(Entity, &mut Transform, &mut Velocity, &MoveToEnemy), With<Minion>>,
     mut bandit_query: Query<(Entity, &Transform), (With<Bandit>, Without<Minion>)>,)
+{
+    let bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
+    for mut minion in minion_query.iter_mut()
     {
-        let bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
-        for mut minion in minion_query.iter_mut(){
 
-            if let Some(bandit) = bandits.iter().find(|t| {
-                t.0 == minion.3.0
-            }){
-                if minion.1.translation.distance(bandit.1.translation) <  150.0
-                { 
-                    change_state::<MoveToEnemy>(&mut commands, minion.0, AttackEnemy(bandit.0, *minion.1)); 
-                    continue;
-                }
-
-                let direction = (bandit.1.translation - minion.1.translation).normalize();
-                minion.2.0 += Vec3{x: direction.x * 25.0, y: direction.y * 25.0, z: 0.0} ;
-            }else {
-                change_state::<FollowPlayer>(&mut commands, minion.0, FollowPlayer);
+        if let Some(bandit) = bandits.iter().find(|t| {
+            t.0 == minion.3.0
+        }){
+            if minion.1.translation.distance(bandit.1.translation) <  75.0
+            { 
+                change_state::<MoveToEnemy>(&mut commands, minion.0, AttackEnemy(bandit.0, *minion.1)); 
+                continue;
             }
+
+            let direction = (bandit.1.translation - minion.1.translation).normalize();
+            minion.2.0 += Vec3{x: direction.x * 25.0, y: direction.y * 25.0, z: 0.0} ;
+        }else {
+            change_state::<FollowPlayer>(&mut commands, minion.0, FollowPlayer);
         }
+    }
 }
 
 
@@ -127,26 +116,27 @@ pub fn attack_enemy(
     mut bandit_query: Query<(Entity, &Transform), (With<Bandit>, Without<Minion>)>,
     mut writer: EventWriter<Attack>,
     time: Res<Time>)
-    {  
-        if minion_query.is_empty() { return; }
-        let bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
-       
-        for mut minion in minion_query.iter_mut(){
-                if let Some(bandit) = bandits.iter().find(|t| t.0 == minion.3.0)
-                {
-                    if minion.4.translation.distance(bandit.1.translation) > 150.0 
-                    { 
-                        change_state::<AttackEnemy>(&mut commands, minion.0, MoveToEnemy(minion.3.0, *bandit.1));
-                        continue;
-                    }
-
-                    if minion.1.last_attacked > time.elapsed_seconds() - 1.5   { continue; }
-                    writer.send(Attack { from:*minion.4 , to: minion.3.0 , damage: minion.2.0 });   
-                    minion.1.last_attacked = time.elapsed_seconds();
-                } else {
-                    change_state::<AttackEnemy>(&mut commands, minion.0, FollowPlayer)
+{  
+    if minion_query.is_empty() { return; }
+    let bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
+    
+    for mut minion in minion_query.iter_mut()
+    {
+            if let Some(bandit) = bandits.iter().find(|t| t.0 == minion.3.0)
+            {
+                if minion.4.translation.distance(bandit.1.translation) > 75.0 
+                { 
+                    change_state::<AttackEnemy>(&mut commands, minion.0, MoveToEnemy(minion.3.0, *bandit.1));
+                    continue;
                 }
-        }
+
+                if minion.1.last_attacked > time.elapsed_seconds() - 1.5   { continue; }
+                writer.send(Attack { from:*minion.4 , to: minion.3.0 , damage: minion.2.0 });   
+                minion.1.last_attacked = time.elapsed_seconds();
+            } else {
+                change_state::<AttackEnemy>(&mut commands, minion.0, FollowPlayer)
+            }
+    }
 }
 
 const ENEMY_RANGE: f32 = 500.0;
@@ -155,26 +145,26 @@ pub fn check_enemy_range(
     mut commands: Commands,
     mut minion_query: Query<(Entity, &mut Transform, &mut Velocity), (With<Minion>, With<FollowPlayer>)>,
     mut bandit_query: Query<(Entity, &Transform), (With<Bandit>, Without<Minion>)>)
+{
+    if bandit_query.is_empty(){return;}
+
+    let mut bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
+    for minion in minion_query.iter_mut()
     {
-        if bandit_query.is_empty(){return;}
-
-        let mut bandits  = bandit_query.iter_mut().collect::<Vec<(Entity, &Transform)>>();
-        for minion in minion_query.iter_mut(){
-            let bandit = get_nearest_entity(&mut bandits, minion.1.translation);
-           
-            if bandit.1.translation.distance(minion.1.translation) > ENEMY_RANGE {
-                continue;
-            }
-
-            change_state::<FollowPlayer>(&mut commands, minion.0, MoveToEnemy(bandit.0, bandit.1))
+        let bandit = get_nearest_entity(&mut bandits, minion.1.translation);
+        
+        if bandit.1.translation.distance(minion.1.translation) > ENEMY_RANGE {
+            continue;
         }
+
+        change_state::<FollowPlayer>(&mut commands, minion.0, MoveToEnemy(bandit.0, bandit.1))
+    }
 }
 
 pub fn minion_follow_player(
     mut minion_query: Query<(&mut Transform, &mut Velocity), (With<Minion>, Without<Player>, With<FollowPlayer>)>,
     mut player_query: Query<&mut Transform, With<Player>>) 
-    {
-
+{
     let player = player_query.single_mut();
     for mut minion in minion_query.iter_mut() {
         if minion.0.translation.distance(player.translation) <  150.0 { continue; }
